requirements.txt
streamlit==1.52.2
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.10.8
pillow==10.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
zipp==3.23.0
xlsxwriter==3.2.9
